pyproject.toml
[project]
name = "DocsToKG"
version = "0.2.0"
description = "DocsToKG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # ============================================================================
    # HTTP & Networking
    # ============================================================================
    "httpx[http2]>=0.28.1",
    "hishel>=0.1.0",
    "certifi>=2025.10.5",
    "pyrate-limiter>=3.9,<4",
    "url-normalize==2.2.*",

    # ============================================================================
    # Data Processing & Storage
    # ============================================================================
    "pyarrow>=21.0.0",
    "numpy>=2.2.6",
    "jsonlines>=4.0.0",
    "libarchive-c>=5.3",

    # ============================================================================
    # Pydantic Ecosystem
    # ============================================================================
    "pydantic>=2.0,<3.0",
    "pydantic-core>=2.41.4",
    "pydantic-settings>=2.11.0",

    # ============================================================================
    # Document Processing
    # ============================================================================
    "docling-core>=2.48.4",

    # ============================================================================
    # Domain-Specific APIs
    # ============================================================================
    "pyalex>=0.18",
    "waybackpy>=3.0.6",

    # ============================================================================
    # Ontology & Knowledge Graphs
    # ============================================================================
    "bioregistry>=0.12.43",
    "arelle>=2.2",
    "rdflib>=7.2.1",

    # ============================================================================
    # Web Scraping
    # ============================================================================
    "trafilatura>=2.0.0",
    "beautifulsoup4>=4.14.2",

    # ============================================================================
    # ML/Embeddings (lazy imports - consider making optional)
    # ============================================================================
    "transformers>=4.57.1",

    # ============================================================================
    # Observability & Reliability
    # ============================================================================
    "prometheus-client>=0.23.1",
    "tenacity>=9.1.2",
    "pybreaker",

    # ============================================================================
    # CLI & User Experience
    # ============================================================================
    "typer>=0.19.2",
    "rich>=14.2.0",
    "tqdm>=4.67.1",

    # ============================================================================
    # Utilities
    # ============================================================================
    "filelock>=3.20.0",
    "platformdirs>=4.5.0",
    "typing-extensions>=4.15.0",
    "pyyaml>=6",
    "jsonschema>=4.25.1",

    # ============================================================================
    # Testing Framework (if tests are in src/)
    # ============================================================================
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-logging>=2015.11.4",
    "hypothesis>=6.141.0",
    "requests>=2.32.5",
    "psutil>=7.1.0",

    # ============================================================================
    # Cloud & Databases
    # ============================================================================
    "boto3",  # for AWS S3 storage
    "psycopg>=3",  # for PostgreSQL
    "psycopg-pool",  # for PostgreSQL connection pooling
    "sqlalchemy>=2.0.44",  # for PostgreSQL via ORM
    "redis>=6.4.0",  # for Redis caching

    # ============================================================================
    # Advanced Analytics
    # ============================================================================
    "polars",
    "duckdb>=1.4.1",
]


[project.optional-dependencies]
gpu = [
    "nvidia-cublas-cu12>=12.9.1.4",
    "nvidia-cuda-cupti-cu12>=12.9.79",
    "nvidia-cuda-nvrtc-cu12>=12.9.86",
    "nvidia-cuda-runtime-cu12>=12.9.79",
    "nvidia-cudnn-cu12>=9.10.2.21",
    "nvidia-cudnn-frontend>=1.15.0",
    "nvidia-cufft-cu12>=11.4.1.4",
    "nvidia-cufile-cu12>=1.14.1.1",
    "nvidia-curand-cu12>=10.3.10.19",
    "nvidia-cusolver-cu12>=11.7.5.82",
    "nvidia-cusparse-cu12>=12.5.10.65",
    "nvidia-cusparselt-cu12>=0.7.1",
    "nvidia-cutlass-dsl>=4.2.1",
    "nvidia-ml-py>=13.580.82",
    "nvidia-nccl-cu12>=2.27.3",
    "nvidia-nvjitlink-cu12>=12.9.86",
    "nvidia-nvshmem-cu12>=3.4.5",
    "nvidia-nvtx-cu12>=12.9.79",
    "cuda-bindings>=12.9.3",
    "cuda-pathfinder>=1.3.1",
    "cuda-python>=12.9.3",
    "cuda-toolkit>=12.9.1",
    "pylibraft-cu12>=25.10.0",
    "rmm-cu12>=25.10.0",
    "torch @ https://download.pytorch.org/whl/cu129/torch-2.8.0%2Bcu129-cp313-cp313-manylinux_2_28_x86_64.whl",
    "torchvision @ https://download.pytorch.org/whl/cu129/torchvision-0.23.0%2Bcu129-cp313-cp313-manylinux_2_28_x86_64.whl",
    "torchaudio @ https://download.pytorch.org/whl/cu129/torchaudio-2.8.0%2Bcu129-cp313-cp313-manylinux_2_28_x86_64.whl",
    "flashinfer-cubin>=0.4.0",
    "apache-tvm-ffi==0.1.0b15",
    "flashinfer-python>=0.4.0",
    "compressed-tensors>=0.11.0",
    "accelerate>=1.10.1",
    "easyocr>=1.7.2",
    "libcuvs-cu12>=25.10.0",
    "libraft-cu12>=25.10.0",
    "librmm-cu12>=25.10.0",
    "cuvs-cu12>=25.10.0",
    "docling>=2.56.1",
    "docling-ibm-models>=3.9.1",
    "openai>=2.3.0",
    "openai-harmony>=0.0.4",
    "openai-whisper>=20250625",
    "pyserini>=1.2.0",
    "transformers>=4.57.1",
    "triton>=3.4.0",
    "sentence-transformers>=5.1.1",
    "keybert>=0.9.0",
    "stanza>=1.11.0",
    "xformers>=0.0.32.post1",
    "xgrammar>=0.1.25",
    "safetensors>=0.6.2",
    "cupy-cuda12x==13.6.0 ; python_version >= '3.13' and python_version < '3.14'",
    "vllm==0.11.0; python_version >= '3.13' and python_version < '3.14' and platform_system == 'Linux' and platform_machine == 'x86_64'",
    "faiss==1.12.0 ; python_version >= '3.13' and python_version < '3.14' and platform_system == 'Linux' and platform_machine == 'x86_64'",
    "ray>=2.50.0",
]
dev = [
    "mypy>=1.11.0",
    "mypy-extensions>=1.0.0",
    "types-requests>=2.32.0.20241016",
    "types-PyYAML>=6.0.12.20240917",
]
docs = [
    "sphinx>=7.3.7",
    "sphinx-autobuild>=2024.4.3",
    "sphinx-autodoc-typehints>=2.2.1",
]

[project.scripts]
docstokg-hybrid-validation = "DocsToKG.HybridSearch.validation:main"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E203"]

[tool.pytest.ini_options]
markers = [
    "real_vectors: tests that require the real-vector fixture (enable with --real-vectors)",
    "scale_vectors: tests that require the large real-vector fixture (enable with --scale-vectors)",
    "integration: marks integration tests that may hit real network services",
    "benchmark: performance benchmarks requiring pytest-benchmark",
]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
strict_equality = true
show_error_codes = true

[tool.uv.pip]
find-links = ["./.wheelhouse"]
